# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghana School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: termId
          in: query
          required: false
          schema:
            type: integer
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for a class on a date (bulk)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAttendanceBody"
      responses:
        "201":
          description: Attendance marked
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary by class and term
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List terminal results
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      summary: Record a student result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResultBody"
      responses:
        "201":
          description: Result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /results/{id}:
    put:
      operationId: updateResult
      tags: [results]
      summary: Update a result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResultBody"
      responses:
        "200":
          description: Updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
    delete:
      operationId: deleteResult
      tags: [results]
      summary: Delete result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /results/bulk:
    post:
      operationId: bulkCreateResults
      tags: [results]
      summary: Bulk enter results for a class/subject/term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateResultsBody"
      responses:
        "201":
          description: Results created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
  /report-cards:
    get:
      operationId: listReportCards
      tags: [results]
      summary: Get report card data (aggregated results per student per term)
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Report cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportCard"
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: ["string", "null"]
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        termId:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, classId, date, status, createdAt]
    MarkAttendanceBody:
      type: object
      properties:
        classId:
          type: integer
        date:
          type: string
          format: date
        termId:
          type: integer
        records:
          type: array
          items:
            type: object
            properties:
              studentId:
                type: integer
              status:
                type: string
                enum: [present, absent, late, excused]
              remarks:
                type: string
            required: [studentId, status]
      required: [classId, date, records]
    AttendanceSummary:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        excusedDays:
          type: integer
        attendanceRate:
          type: number
      required: [studentId, studentName, classId, totalDays, presentDays, absentDays, lateDays, excusedDays, attendanceRate]
...
    Result:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        termId:
          type: integer
        termName:
          type: ["string", "null"]
        classScore:
          type: ["number", "null"]
        examScore:
          type: ["number", "null"]
        totalScore:
          type: ["number", "null"]
        grade:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        position:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, classId, subjectId, termId, createdAt]
    CreateResultBody:
      type: object
      properties:
        studentId:
          type: integer
        classId:
          type: integer
        subjectId:
          type: integer
        termId:
          type: integer
        classScore:
          type: number
        examScore:
          type: number
        remarks:
          type: string
      required: [studentId, classId, subjectId, termId]
    UpdateResultBody:
      type: object
      properties:
        classScore:
          type: number
        examScore:
          type: number
        remarks:
          type: string
    BulkCreateResultsBody:
      type: object
      properties:
        classId:
          type: integer
        subjectId:
          type: integer
        termId:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              studentId:
                type: integer
              classScore:
                type: number
              examScore:
                type: number
              remarks:
                type: string
            required: [studentId]
      required: [classId, subjectId, termId, results]
    ReportCard:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        studentId_code:
          type: string
        className:
          type: string
        termName:
          type: string
        academicYearName:
          type: string
        results:
          type: array
          items:
            type: object
            properties:
              subjectName:
                type: string
              classScore:
                type: ["number", "null"]
              examScore:
                type: ["number", "null"]
              totalScore:
                type: ["number", "null"]
              grade:
                type: ["string", "null"]
              remarks:
                type: ["string", "null"]
              position:
                type: ["integer", "null"]
            required: [subjectName]
        totalScore:
          type: ["number", "null"]
        averageScore:
          type: ["number", "null"]
        overallGrade:
          type: ["string", "null"]
        classPosition:
          type: ["integer", "null"]
        totalStudents:
          type: ["integer", "null"]
        attendanceSummary:
          type: object
          properties:
            totalDays:
              type: integer
            presentDays:
              type: integer
            attendanceRate:
              type: number
          required: [totalDays, presentDays, attendanceRate]
        conductRemarks:
          type: ["string", "null"]
        teacherRemarks:
          type: ["string", "null"]
      required: [studentId, studentName, studentId_code, className, termName, academicYearName, results]
...